pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "precis"
version = "0.1.0"
description = "A framework for automating your media diet"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["rss", "llm", "ai"]

authors = [
  {name = "Leo Qin", email = "me@leozqin.me" }
]
maintainers = [
  {name = "Leo Qin", email = "me@leozqin.me" }
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Office/Business :: News/Diary",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic",
    "feedparser",
    "aenum",
    "ruamel.yaml",
    "ruamel.yaml.string",
    "fastapi==0.110.1",
    "uvicorn[standard]",
    "tinydb",
    "fastapi-utils",
    "jinja2",
    "html2text",
    "lxml_html_clean",
    "requests",
    "readabilipy",
    "readability-lxml",
    "markdown2[all]",
    "simplematrixbotlib",
    "slack_sdk",
    "ollama",
    "playwright",
    "openai",
    "python-multipart"
]

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Reports" = "https://github.com/pypa/sampleproject/issues"
"Funding" = "https://donate.pypi.org"
"Say Thanks!" = "http://saythanks.io/to/example"
"Source" = "https://github.com/pypa/sampleproject/"

# The following would provide a command line executable called `sample`
# which executes the function `main` from this package when invoked.
[project.scripts]
precis = "app.cli:cli"


# This is configuration specific to the `setuptools` build backend.
# If you are using a different build backend, you will need to change this.
[tool.setuptools]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {"sample" = ["*.dat"]}
packages = ["app"]
